fixed excess recommendations regarding entropy
When there would be only 2 possible solutions left but  > 2 guessable words remaining, it would recommend them all but because entropy at this stage = 1.0 it is more efficient to simply recommend those 2 possible solutions. Entropy cannot be reduced any more than via guessing 1/2 of the solution words since it would either be the solution or now we know the solution is the other option.